end of div and sub

+++ main/header.h
@@ -2,8 +2,8 @@
 #define HEADER
 double MySum(int, int);
 double MyMul(int, int);
-double MySub(int, int);
-double MyDiv(int, int);
+double MySub(double, double);
+double MyDiv(double, double);
 
 #endif // HEADER
